Move KV helpers test to corresponding file
